Add tooltip special keywords
The keyword `all` will include tooltips for all columns in the datasource (except for the processed coordinate columns, which don't render well in tooltips); the keyword `meta` will include all columns in the data source except for the unprocessed, original location data and the x and y centroids; `point` will include the centroids; `raw_data` will include the unprocessed location data. If none of these optios are used, tooltips need to be fed a list of tuples, where the first element of each tuple is the title of the tooltip element, and the second element is the name of the relevant column in the data source, preceeded by `@`.